Add delete file rules
