+++ src/routes/arena.ts
@@ -5,7 +5,6 @@ import * as fs from "fs/promises";
 import * as utils from "../helpers/utils";
 import authenticate, { JwtServerPayload } from "../middlewares/authenticate";
 import * as hasura from "../helpers/hasura"
-import { join } from "path";
 
 
 const router = express.Router();
@@ -417,21 +416,7 @@ router.post("/finish", async (req, res) => {
       console.log("Files uploaded!")
 
       try {
-        const deleteFile = async function deleteAllFilesInDir(directoryPath: string) {
-          const files = await fs.readdir(directoryPath);
-          await Promise.all(files.map(async (file) => {
-            const filePath = join(directoryPath, file);
-            const stats = await fs.stat(filePath);
-            if (stats.isDirectory()) {
-              await deleteAllFilesInDir(filePath);
-            } else {
-              await fs.unlink(filePath);
-            }
-          }
-          ));
-          await fs.rmdir(directoryPath);
-        }
-        await deleteFile(`${base_directory}/${contest_name}/arena/${room_id}`);
+        await utils.deleteAllFilesInDir(`${base_directory}/${contest_name}/arena/${room_id}`);
 
       } catch (err) {
         return res.status(500).send("500 Internal Server Error: Delete files failed. " + err);
